feat(ghg): abstract generic survey table
