+++ frontend/src/components/green_house_gaz/energy/Cooking.vue
@@ -12,9 +12,13 @@
 </template>
 
 <script lang="ts">
-import BaselineEndlineWrapper, {
+import BaselineEndlineWrapper from "@/components/green_house_gaz/generic/BaselineEndlineWrapper.vue";
+import {
+  ensureSurveyTableHeaders,
   SurveyTableHeader,
-} from "@/components/green_house_gaz/generic/BaselineEndlineWrapper.vue";
+  surveyTableHeaderCO2,
+  surveyTableHeaderIncrements,
+} from "@/components/green_house_gaz/generic/surveyTableHeader";
 import SurveyItemTitle from "@/components/green_house_gaz/SurveyItemTitle.vue";
 
 import { ReferenceItemInterface } from "@/store/GhgReferenceModule";
@@ -33,7 +37,6 @@ import {
   SurveyItem,
   SurveyResult,
 } from "@/store/GhgInterface.vue";
-import { get as _get } from "lodash";
 
 import { GreenHouseGaz, Survey } from "@/store/GhgInterface.vue";
 import { GHGfNRB } from "@/store/GHGReferencefNRB";
@@ -322,18 +325,6 @@ export default class Cooking extends Vue {
     };
   }
 
-  n2sFormatter(n: number): string {
-    // https://stackoverflow.com/a/30686832
-    let s = "";
-    if (!n) s = "a";
-    else
-      while (n) {
-        s = String.fromCharCode(97 + (n % 26)) + s;
-        n = Math.floor(n / 26);
-      }
-    return s;
-  }
-
   resetSurveyInput(localInput: EnergyCookingItemInput): EnergyCookingItemInput {
     localInput.fuelType = "NO_ACCESS";
     delete localInput.numberOfCookstove; // percentage of percentage
@@ -364,37 +355,7 @@ export default class Cooking extends Vue {
     // public get headers(): any {
     const countryCode = this.countryCode;
     return [
-      {
-        text: "#", // unique name === dropdown of existant facilities
-        value: "increment",
-        type: "number",
-        hideFooterContent: false,
-        baselineOnly: true,
-        formatter: this.n2sFormatter,
-      },
-      {
-        text: "#", // unique name === dropdown of existant facilities
-        value: "originIncrement",
-        endlineOnly: true,
-        type: "number",
-        hideFooterContent: false,
-        formatter: (
-          v: number,
-          _: unknown,
-          item: SurveyItem,
-          items: SurveyItem[]
-        ) => {
-          const increment: number = _get(item, "increment");
-          const increments = items
-            .filter((item: SurveyItem) => item.originIncrement === v)
-            .map((item: SurveyItem) => item.increment);
-          const indexOf = increments.indexOf(increment);
-          return `${this.n2sFormatter(v)}${"'".repeat(indexOf)}`;
-        },
-        formatterOrigin: (v: number) => {
-          return `${this.n2sFormatter(v)}`;
-        },
-      },
+      ...surveyTableHeaderIncrements,
       {
         text: "Cookstove",
         value: "input.cookstove",
@@ -610,92 +571,12 @@ export default class Cooking extends Vue {
           });
         },
       },
-      {
-        text: "Total CO2 Emissions (tCO2e/year)",
-        value: "computed.totalCO2Emission",
-        hideFooterContent: false,
-        formatter: (v: number, { ...args }) => {
-          return formatNumber(v, { suffix: args.suffix });
-        },
-        computeResults: true,
-        type: "number",
-        disabled: true,
-      },
-      {
-        text: "Change in Emissions",
-        value: "computed.changeInEmission",
-        type: "number",
-        hideFooterContent: false,
-        disable: true,
-        readonly: true,
-        endlineOnly: true,
-        formatter: (v: number) => {
-          return formatNumber(v, {
-            style: "percent",
-            signDisplay: "exceptZero",
-          });
-        },
-        classFormatter: (v: number): string => {
-          const classes: string[] = [];
-          v > 0 ? classes.push("item-positive") : void 0;
-          v < 0 ? classes.push("item-negative") : void 0;
-          v === 0 ? classes.push("bold-table-content") : void 0;
-          return classes.join(" ");
-        },
-      },
-      {
-        text: "",
-        value: "actions",
-        hidden: true,
-        hideFooterContent: false,
-        width: "140px",
-      },
+      ...surveyTableHeaderCO2,
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ].map((item: any) => {
-      const [category, key] = item.value.split(".");
-      const isInput = item?.isInput ?? category === "input";
-      return {
-        align: "start",
-        sortable: false,
-        hideFooterContent: item.hideFooterContent ?? true,
-        label: item.text, // for form-item-component
-        key, // for form-item-component
-        isInput,
-        category, // input or computed,
-        formatter: (value: unknown) => value,
-        classFormatter: () => "",
-        options: (() => {
-          if (item.options) {
-            return item.options;
-          }
-          const items = item?.items;
-          if (typeof items === "string") {
-            // items should not be string.
-            return [];
-          }
-          return (
-            items?.map((item: string | SelectCustom<string>) => {
-              if (typeof item === "string") {
-                return { text: item, value: item };
-              }
-              return {
-                text: item?.text,
-                value: item?._id,
-              };
-            }) ?? undefined
-          );
-        })(),
-        ...item,
-      } as SurveyTableHeader;
-    });
+    ].map(ensureSurveyTableHeaders);
   }
 
   items!: GHGfNRB[];
-}
-
-export interface SelectCustom<V> {
-  text: string;
-  _id: V;
 }
 
 export interface EnergyCookingItemInput
