+++ frontend/src/components/green_house_gaz/energy/Facilities.vue
@@ -13,9 +13,7 @@
 </template>
 
 <script lang="ts">
-import BaselineEndlineWrapper, {
-  SurveyTableHeader,
-} from "@/components/green_house_gaz/generic/BaselineEndlineWrapper.vue";
+import BaselineEndlineWrapper from "@/components/green_house_gaz/generic/BaselineEndlineWrapper.vue";
 import SurveyItemTitle from "@/components/green_house_gaz/SurveyItemTitle.vue";
 
 import { ReferenceItemInterface } from "@/store/GhgReferenceModule";
@@ -27,6 +25,12 @@ import {
   solarInputsProducedPer,
 } from "@/components/green_house_gaz/energy/solarInputs";
 
+import {
+  ensureSurveyTableHeaders,
+  SurveyTableHeader,
+  surveyTableHeaderCO2,
+  surveyTableHeaderIncrements,
+} from "@/components/green_house_gaz/generic/surveyTableHeader";
 import { formatNumber } from "@/plugins/filters";
 import {
   DieselItem,
@@ -36,7 +40,6 @@ import {
   SurveyItem,
   SurveyResult,
 } from "@/store/GhgInterface.vue";
-import { get as _get } from "lodash";
 
 import {
   ElectricFuel,
@@ -147,18 +150,6 @@ export default class Facilities extends Vue {
     };
   }
 
-  n2sFormatter(n: number): string {
-    // https://stackoverflow.com/a/30686832
-    let s = "";
-    if (!n) s = "a";
-    else
-      while (n) {
-        s = String.fromCharCode(97 + (n % 26)) + s;
-        n = Math.floor(n / 26);
-      }
-    return s;
-  }
-
   resetSurveyInput(
     localInput: EnergyFacilityItemInput
   ): EnergyFacilityItemInput {
@@ -186,40 +177,8 @@ export default class Facilities extends Vue {
 
   // should be a getter so it may be reactive for fuelTypes
   public get headers(): SurveyTableHeader[] {
-    // public get headers(): any {
-    const countryCode = this.countryCode;
     return [
-      {
-        text: "#", // unique name === dropdown of existant facilities
-        value: "increment",
-        type: "number",
-        hideFooterContent: false,
-        baselineOnly: true,
-        formatter: this.n2sFormatter,
-      },
-      {
-        text: "#", // unique name === dropdown of existant facilities
-        value: "originIncrement",
-        endlineOnly: true,
-        type: "number",
-        hideFooterContent: false,
-        formatter: (
-          v: number,
-          _: unknown,
-          item: SurveyItem,
-          items: SurveyItem[]
-        ) => {
-          const increment: number = _get(item, "increment");
-          const increments = items
-            .filter((item: SurveyItem) => item.originIncrement === v)
-            .map((item: SurveyItem) => item.increment);
-          const indexOf = increments.indexOf(increment);
-          return `${this.n2sFormatter(v)}${"'".repeat(indexOf)}`;
-        },
-        formatterOrigin: (v: number) => {
-          return `${this.n2sFormatter(v)}`;
-        },
-      },
+      ...surveyTableHeaderIncrements,
       {
         text: "Name",
         value: "input.name",
@@ -280,7 +239,7 @@ export default class Facilities extends Vue {
         type: "number",
       },
       // end of national grid\
-      ...solarInputsProducedPer("Year", countryCode, this.project.solar),
+      ...solarInputsProducedPer("Year", this.countryCode, this.project.solar),
       {
         text: "Total (kWh/yr)",
         value: "computed.totalPower",
@@ -294,93 +253,11 @@ export default class Facilities extends Vue {
         type: "number",
         disabled: true,
       },
+      ...surveyTableHeaderCO2,
 
-      {
-        text: "Total CO2 Emissions (tCO2e/yr)",
-        value: "computed.totalCO2Emission",
-        hideFooterContent: false,
-        formatter: (v: number, { ...args }) => {
-          return formatNumber(v, { suffix: args.suffix });
-        },
-        computeResults: true,
-        type: "number",
-        disabled: true,
-      },
-      {
-        text: "Change in Emissions",
-        value: "computed.changeInEmission",
-        type: "number",
-        hideFooterContent: false,
-        disable: true,
-        readonly: true,
-        endlineOnly: true,
-        formatter: (v: number) => {
-          return formatNumber(v, {
-            style: "percent",
-            signDisplay: "exceptZero",
-          });
-        },
-        classFormatter: (v: number): string => {
-          const classes: string[] = [];
-          v > 0 ? classes.push("item-positive") : void 0;
-          v < 0 ? classes.push("item-negative") : void 0;
-          v === 0 ? classes.push("bold-table-content") : void 0;
-          return classes.join(" ");
-        },
-      },
-      {
-        text: "",
-        value: "actions",
-        hidden: true,
-        hideFooterContent: false,
-        width: "140px",
-      },
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ].map((item: any) => {
-      // todo: externalize this map function in the generic component to make it DRY
-      const [category, key] = item.value.split(".");
-      const isInput = item?.isInput ?? category === "input";
-      return {
-        align: "start",
-        sortable: false,
-        hideFooterContent: item.hideFooterContent ?? true,
-        hideInput: item.hideInput ?? false,
-        label: item.text, // for form-item-component
-        key, // for form-item-component
-        isInput,
-        category, // input or computed,
-        formatter: (value: unknown) => value,
-        classFormatter: () => "",
-        options: (() => {
-          if (item.options) {
-            return item.options;
-          }
-          const items = item?.items;
-          if (typeof items === "string") {
-            // items should not be string.
-            return [];
-          }
-          return (
-            items?.map((item: string | SelectCustom<string>) => {
-              if (typeof item === "string") {
-                return { text: item, value: item };
-              }
-              return {
-                text: item?.text,
-                value: item?._id,
-              };
-            }) ?? undefined
-          );
-        })(),
-        ...item,
-      } as SurveyTableHeader;
-    });
+    ].map(ensureSurveyTableHeaders);
   }
-}
-
-export interface SelectCustom<V> {
-  text: string;
-  _id: V;
 }
 
 export interface EnergyFacilityItemInput
