+++ frontend/src/components/green_house_gaz/generic/BaselineCard.vue
@@ -60,9 +60,9 @@
 </template>
 
 <script lang="ts">
-import { SurveyTableHeader } from "@/components/green_house_gaz/generic/BaselineEndlineWrapper.vue";
 import InstancePieChart from "@/components/green_house_gaz/generic/InstancePieChart.vue";
 import InstanceTable from "@/components/green_house_gaz/generic/InstanceTable.vue";
+import { SurveyTableHeader } from "@/components/green_house_gaz/generic/surveyTableHeader";
 import { SurveyInput, SurveyItem } from "@/store/GhgInterface.vue";
 import Vue from "vue";
 import "vue-class-component/hooks";
